backend/requirements.txt
fastapi==0.103.0
uvicorn[standard]==0.23.2
pydantic==2.4.2
pydantic-core==2.10.1
pydantic-settings==2.0.3
cloudinary==1.36.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
sentence-transformers==2.7.0
numpy==1.26.4
webcolors==1.13
colormath==3.0.0
torch==2.8.0+cpu
httpx==0.24.1
--extra-index-url https://download.pytorch.org/whl/cpu
